refactor: seperatd the DBManeger into 2 calasses 1 that handle with anything has to do with person and the other for inrtel
